Move test_deserialize_empty_collections to cql_types.rs
